chips: nrf52: split out peripherals to a file
