+++ chips/nrf52840/src/lib.rs
@@ -6,8 +6,8 @@
 
 pub use nrf52::{
     acomp, adc, aes, ble_radio, chip, clock, constants, crt1, ficr, i2c, ieee802154_radio, init,
-    nvmc, peripheral_interrupts as base_interrupts, pinmux, power, ppi, pwm, rtc, spi, temperature,
-    timer, trng, uart, uicr, usbd,
+    nvmc, peripheral_interrupts as base_interrupts, peripherals, pinmux, power, ppi, pwm, rtc, spi,
+    temperature, timer, trng, uart, uicr, usbd,
 };
 
 pub use nrf52::nrf52_default_peripherals_component_static;
